Fix: Eclair max fee percent needs to be integer value

+++ src/BTCPayServer.Lightning.Eclair/Models/PayInvoiceRequest.cs
@@ -5,7 +5,7 @@ namespace BTCPayServer.Lightning.Eclair.Models
         public string Invoice { get; set; }
         public int? AmountMsat { get; set; }
         public int? MaxAttempts { get; set; }
-        public double? MaxFeePct { get; set; }
+        public int? MaxFeePct { get; set; }
         public long? MaxFeeFlatSat { get; set; }
     }
 }
